solve the problem of Happens-before Link:
A Happens before Link exists between
all synchronized or volatile write operations and
all synchronized or volatile read operations that follow.

Synchronization guarantees the exclusive execution of a block of code.
Visibility guarantees the consistency of the variables, means a read should return the value set by the last write.